chore: move action button
