+++ src/extension/options-page/DashboardDialogs/Persona.tsx
@@ -4,7 +4,7 @@ import { DialogContentItem } from './DialogBase'
 import { TextField, Typography, InputBase, makeStyles } from '@material-ui/core'
 import { Link, useHistory } from 'react-router-dom'
 import BackupRestoreTab, { BackupRestoreTabProps } from '../DashboardComponents/BackupRestoreTab'
-import ActionButton from '../../../components/Dashboard/ActionButton'
+import ActionButton from '../DashboardComponents/ActionButton'
 import { ECKeyIdentifier, ProfileIdentifier } from '../../../database/type'
 import Services from '../../service'
 import { Persona } from '../../../database'
