+++ src/extension/options-page/DashboardInitSteps/Step1R.tsx
@@ -5,7 +5,7 @@ import { geti18nString } from '../../../utils/i18n'
 import { useDragAndDrop } from '../../../utils/hooks/useDragAndDrop'
 import { Link, useHistory } from 'react-router-dom'
 import BackupRestoreTab, { BackupRestoreTabProps } from '../DashboardComponents/BackupRestoreTab'
-import ActionButton from '../../../components/Dashboard/ActionButton'
+import ActionButton from '../DashboardComponents/ActionButton'
 import { decompressBackupFile } from '../../../utils/type-transform/BackupFileShortRepresentation'
 import Services from '../../service'
 import { DialogRouter } from '../DashboardDialogs/DialogBase'
