+++ src/extension/options-page/DashboardInitSteps/Step2R.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useState } from 'react'
 import StepBase from './StepBase'
 import { Table, TableBody, TableRow, TableCell, styled } from '@material-ui/core'
 import { geti18nString } from '../../../utils/i18n'
-import ActionButton from '../../../components/Dashboard/ActionButton'
+import ActionButton from '../DashboardComponents/ActionButton'
 
 const header = 'Restoration Successful'
 const subheader = 'Restored from a backup at 2020-01-01 09:41 (UTC+0).'
