+++ src/extension/options-page/DashboardInitSteps/Step2S.tsx
@@ -3,7 +3,7 @@ import StepBase from './StepBase'
 import { Typography } from '@material-ui/core'
 import { geti18nString } from '../../../utils/i18n'
 import { Link } from 'react-router-dom'
-import ActionButton from '../../../components/Dashboard/ActionButton'
+import ActionButton from '../DashboardComponents/ActionButton'
 import useQueryParams from '../../../utils/hooks/useQueryParams'
 import Services from '../../service'
 import { ECKeyIdentifier } from '../../../database/type'
